increment stripes to v5 react-router to v5.2
Increment `@folio/stripes` to `v5` in concert with incrementing
`react-router` to `v5.2` (which was also moved from a direct dep to a
peer so we can be sure to receive the correct context from
stripes-core). Corresponding dev-deps on other stripes-* libraries are
updated as well.

Remove `babel-polyfill`, which has been deprecated, in favor of
`core-js` and `regenerator-runtime`
(https://babeljs.io/docs/en/babel-polyfill).