UIEUS-343 bump react-dropzone to v10
Bump `react-dropzone` from `v9` to `v10`. It is still far out of date
(current is `v14`) but the only breaking change in `v10` is dropping
support for `react` `v16` so this seems like the smallest/safest change.
`v11` and up contain more extensive changes that need to be carefully
evaluated.

Refs UIEUS-343